remove `trace_id` entirely, use meta-maps in more places
The direct use of span_id+trace_id in OTEL is a bad idea, and it makes
little sense in other collectors. In the end, `span` should be thought
of as an opaque key only the collector understands, and a meta-map
should be carried around for the collector to stash additional info such
as a (8B, 16B, ...) trace ID or other information.